Moved FCM Notify fuction to lambda
